Move ire components to packages
